backend/requirements.txt
Flask
Flask-CORS
Flask-PyMongo
Flask-JWT-Extended
pydantic
python-dotenv
gunicorn